lib/requirements.txt
google-cloud-storage
google-cloud-vision
firebase-admin